combined newsletter and sessions into one table on analytics homepage
